move numeric transform primitives
