+++ featuretools/primitives/standard/transform/__init__.py
@@ -1,24 +1,14 @@
 # flake8: noqa
-from featuretools.primitives.standard.transform.absolute import Absolute
 from featuretools.primitives.standard.transform.binary import *
-from featuretools.primitives.standard.transform.cosine import Cosine
 from featuretools.primitives.standard.transform.cumulative import *
 from featuretools.primitives.standard.transform.datetime import *
-from featuretools.primitives.standard.transform.diff import Diff
 from featuretools.primitives.standard.transform.email import *
 from featuretools.primitives.standard.transform.exponential import *
 from featuretools.primitives.standard.transform.is_in import IsIn
 from featuretools.primitives.standard.transform.is_null import IsNull
 from featuretools.primitives.standard.transform.latlong import *
 from featuretools.primitives.standard.transform.natural_language import *
-from featuretools.primitives.standard.transform.natural_logarithm import (
-    NaturalLogarithm,
-)
-from featuretools.primitives.standard.transform.negate import Negate
 from featuretools.primitives.standard.transform.not_primitive import Not
-from featuretools.primitives.standard.transform.percentile import Percentile
-from featuretools.primitives.standard.transform.sine import Sine
-from featuretools.primitives.standard.transform.square_root import SquareRoot
-from featuretools.primitives.standard.transform.tangent import Tangent
+from featuretools.primitives.standard.transform.numeric import *
 from featuretools.primitives.standard.transform.time_series import *
 from featuretools.primitives.standard.transform.url import *
