+++ featuretools/primitives/standard/transform/numeric/__init__.py
@@ -1,0 +1,11 @@
+from featuretools.primitives.standard.transform.numeric.absolute import Absolute
+from featuretools.primitives.standard.transform.numeric.cosine import Cosine
+from featuretools.primitives.standard.transform.numeric.diff import Diff
+from featuretools.primitives.standard.transform.numeric.natural_logarithm import (
+    NaturalLogarithm,
+)
+from featuretools.primitives.standard.transform.numeric.negate import Negate
+from featuretools.primitives.standard.transform.numeric.percentile import Percentile
+from featuretools.primitives.standard.transform.numeric.sine import Sine
+from featuretools.primitives.standard.transform.numeric.square_root import SquareRoot
+from featuretools.primitives.standard.transform.numeric.tangent import Tangent
