+++ featuretools/primitives/standard/transform/numeric/cosine.py
