+++ featuretools/primitives/standard/transform/numeric/diff.py
