+++ featuretools/primitives/standard/transform/numeric/negate.py
