+++ featuretools/primitives/standard/transform/numeric/sine.py
